updateAndRefactoring
