+++ src/main/java/ru/skypro/homework/dto/Ads.java
@@ -5,17 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import ru.skypro.homework.enitities.Ad;
 
 import java.util.List;
 
 @Getter
 @AllArgsConstructor
-@NoArgsConstructor
 @Builder
 public class Ads {
     private int count;
-
     private List<Ad> results;
-
 }
