+++ src/main/java/ru/skypro/homework/dto/Comment.java
@@ -1,0 +1,20 @@
+package ru.skypro.homework.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Getter
+@AllArgsConstructor
+@Builder
+public class Comment {
+    private int pk;
+    private int author;
+    private String authorImage;
+    private String authorFirstName;
+    private long createdAt;
+    private String text;
+}
