+++ src/main/java/ru/skypro/homework/dto/Comments.java
@@ -4,17 +4,14 @@ package ru.skypro.homework.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
-import ru.skypro.homework.enitities.Comment;
 
 import java.util.List;
 
 @Getter
 @AllArgsConstructor
-@NoArgsConstructor
 @Builder
 public class Comments {
     private int count;
-    private List<Comment> result;
+    private List<Comment> results;
 
 }
