+++ src/main/java/ru/skypro/homework/dto/CreateOrUpdateAd.java
@@ -9,7 +9,6 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class CreateOrUpdateAd {
