+++ src/main/java/ru/skypro/homework/dto/ExtendedAd.java
@@ -1,4 +1,4 @@
-package ru.skypro.homework.enitities;
+package ru.skypro.homework.dto;
 
 
 import lombok.AllArgsConstructor;
@@ -6,21 +6,12 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Table;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@Table(name = "Extended_Ad")
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class ExtendedAd {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+
     private int pk;
 
     private String authorFirstName;
