--- src/main/java/ru/skypro/homework/enitities/Ad.java
@@ -1,29 +1,0 @@
-package ru.skypro.homework.enitities;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Ad {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int pk;
-
-    private int author;
-    private String mage;
-    private int price;
-    private String title;
-}
