+++ src/main/java/ru/skypro/homework/enitities/Advertisement.java
@@ -1,0 +1,33 @@
+package ru.skypro.homework.enitities;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import javax.validation.constraints.Size;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Advertisement {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(name = "user_id")
+    private int userId;
+
+    @Size(min = 0, max = 10000000, message = "от нуля до 10000000")
+    private int price;
+    private String image;
+    private String title;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false, updatable = false, insertable = false)
+    User user;
+
+}
