+++ src/main/java/ru/skypro/homework/enitities/User.java
@@ -1,16 +1,22 @@
-package ru.skypro.homework.dto;
+package ru.skypro.homework.enitities;
 
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
+import ru.skypro.homework.dto.Role;
+
+import javax.persistence.*;
 
 
+@Entity
+@Table(name = "users")
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class User {
-
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
     private String email;
